use edi database for gwr addresses
